Make smart computer moves faster and efficient
Reordered the lines in which the code is executed. Did away with the
'axis' array which is no longer needed after reordering. Before picking
from adjacents, it is ensured that the axis is determined with the help
of the lastOccupied array if its length is 2. It is ensured that any
non-relevant cells, meaning cells outside of targetted axis, are not in the adjacents array before its last item is stored in lastAdjacent const. Only adjacent cells in the determined axis are kept.